+++ src/thederpgamer/decor/element/blocks/decor/TextProjector.java
@@ -11,6 +11,7 @@ import thederpgamer.decor.element.blocks.ActivationInterface;
 import thederpgamer.decor.element.blocks.Block;
 import thederpgamer.decor.gui.panel.textprojector.TextProjectorConfigDialog;
 import thederpgamer.decor.manager.ResourceManager;
+import thederpgamer.decor.utils.BlockIconUtils;
 
 /**
  * <Description>
@@ -26,19 +27,16 @@ public class TextProjector extends Block implements ActivationInterface {
 
 	@Override
 	public void initialize() {
-		if (GraphicsContext.initialized) {
+		if(GraphicsContext.initialized) {
 			try {
+				BlockIconUtils.createBlockIcon(blockInfo);
 				blockInfo.setTextureId(ElementKeyMap.getInfo(ElementKeyMap.TEXT_BOX).getTextureIds());
-				blockInfo.setTextureId(
-						0, (short) ResourceManager.getTexture("text-projector-front").getTextureId());
-				blockInfo.setBuildIconNum(ResourceManager.getTexture("text-projector-icon").getTextureId());
-
-			} catch (Exception ignored) {
+				blockInfo.setTextureId(0, (short) ResourceManager.getTexture("text-projector-front").getTextureId());
+			} catch(Exception exception) {
+				exception.printStackTrace();
 			}
 		}
-		blockInfo.setDescription(
-				"A block used to project text at a specified location, scale, and rotation.\n"
-				+ "Check the building quick reference menu for a detailed usage guide.");
+		blockInfo.setDescription("A block used to project text at a specified location, scale, and rotation.\nCheck the building quick reference menu for a detailed usage guide.");
 		blockInfo.setInRecipe(true);
 		blockInfo.setShoppable(true);
 		blockInfo.setCanActivate(true);
@@ -55,12 +53,7 @@ public class TextProjector extends Block implements ActivationInterface {
 		ElementKeyMap.getInfo(666).controlling.add(getId());
 		ElementKeyMap.getInfo(399).controlling.add(getId());
 
-		BlockConfig.addRecipe(
-				blockInfo,
-				ElementKeyMap.getInfo(ElementKeyMap.TEXT_BOX).getProducedInFactoryType(),
-				(int) ElementKeyMap.getInfo(ElementKeyMap.TEXT_BOX).getFactoryBakeTime(),
-				new FactoryResource(1, ElementKeyMap.TEXT_BOX),
-				new FactoryResource(50, (short) 440));
+		BlockConfig.addRecipe(blockInfo, ElementKeyMap.getInfo(ElementKeyMap.TEXT_BOX).getProducedInFactoryType(), (int) ElementKeyMap.getInfo(ElementKeyMap.TEXT_BOX).getFactoryBakeTime(), new FactoryResource(1, ElementKeyMap.TEXT_BOX), new FactoryResource(50, (short) 440));
 		BlockConfig.add(blockInfo);
 	}
 
@@ -70,15 +63,9 @@ public class TextProjector extends Block implements ActivationInterface {
 		configDialog.setSegmentPiece(event.getSegmentPiece());
 		configDialog.activate();
 		event.getSegmentPiece().setActive(!event.getSegmentPiece().isActive());
-		if (GameClient.getClientState() != null)
-			GameClient.getClientState()
-			          .getGlobalGameControlManager()
-			          .getIngameControlManager()
-			          .getPlayerGameControlManager()
-			          .getPlayerIntercationManager()
-			          .suspend(true);
+		if(GameClient.getClientState() != null) GameClient.getClientState().getGlobalGameControlManager().getIngameControlManager().getPlayerGameControlManager().getPlayerIntercationManager().suspend(true);
 	}
 
 	@Override
-	public void onLogicActivation(SegmentPieceActivateEvent event) {}
+	public void onLogicActivation(SegmentPieceActivateEvent event) { }
 }
